+++ Tests/Get-BMVariable.Tests.ps1
@@ -31,8 +31,9 @@ BeforeAll {
         )
 
         New-BMEnvironment -Session $script:session -Name $Named -ErrorAction Ignore
-        Enable-BMEnvironment -Session $script:session -Name $Named
-        Get-BMVariable -Session $script:session -EnvironmentName $Named | Remove-BMVariable -Session $script:session -EnvironmentName $Named
+        $Named | Enable-BMEnvironment -Session $script:session
+        Get-BMVariable -Session $script:session -Environment $Named |
+            Remove-BMVariable -Session $script:session -Environment $Named
     }
 
     function GivenServer
@@ -42,7 +43,9 @@ BeforeAll {
             [string]$Named
         )
 
-        Get-BMServer -Session $script:session -Name $Named -ErrorAction Ignore | Remove-BMServer -Session $script:session
+        $Named |
+            Get-BMServer -Session $script:session -ErrorAction Ignore |
+            Remove-BMServer -Session $script:session
         New-BMServer -Session $script:session -Name $Named -Local
     }
 
@@ -53,7 +56,9 @@ BeforeAll {
             [string]$Named
         )
 
-        Get-BMServerRole -Session $script:session -Name $Named -ErrorAction Ignore | Remove-BMServerRole -Session $script:session
+        $Named |
+            Get-BMServerRole -Session $script:session -ErrorAction Ignore |
+            Remove-BMServerRole -Session $script:session
         New-BMServerRole -Session $script:session -Name $Named
     }
 
@@ -158,7 +163,7 @@ BeforeAll {
 
         if( $Named )
         {
-            $optionalParams['Name'] = $Named
+            $optionalParams['Variable'] = $Named
         }
 
         if( $ValueOnly )
@@ -168,27 +173,27 @@ BeforeAll {
 
         if( $ForApplication )
         {
-            $optionalParams['ApplicationName'] = $ForApplication
+            $optionalParams['Application'] = $ForApplication
         }
 
         if( $ForApplicationGroup )
         {
-            $optionalParams['ApplicationGroupName'] = $ForApplicationGroup
+            $optionalParams['ApplicationGroup'] = $ForApplicationGroup
         }
 
         if( $ForEnvironment )
         {
-            $optionalParams['EnvironmentName'] = $ForEnvironment
+            $optionalParams['Environment'] = $ForEnvironment
         }
 
         if( $ForServer )
         {
-            $optionalParams['ServerName'] = $ForServer
+            $optionalParams['Server'] = $ForServer
         }
 
         if( $ForServerRole )
         {
-            $optionalParams['ServerRoleName'] = $ForServerRole
+            $optionalParams['ServerRole'] = $ForServerRole
         }
 
         $script:result = Get-BMVariable -Session $script:session @optionalParams
@@ -211,7 +216,7 @@ Describe 'Get-BMVariable' {
         ThenNoErrorWritten
     }
 
-    It 'should return variable when searchging by name' {
+    It 'should return variable when searching by name' {
         GivenVariable 'Fubar' -WithValue 'Snafu'
         GivenVariable 'Snafu' -WithValue 'Fubar'
         WhenGettingVariable 'Fubar'
@@ -223,7 +228,7 @@ Describe 'Get-BMVariable' {
         It 'should fail' {
             WhenGettingVariable 'IDONOTEXIST!!!!!!' -ErrorAction SilentlyContinue
             ThenNothingReturned
-            ThenError 'does\ not\ exist'
+            ThenError -AtIndex 0 'does\ not\ exist'
         }
 
         It 'should ignore errors' {
@@ -262,7 +267,8 @@ Describe 'Get-BMVariable' {
         ThenVariableValuesReturned 'Fubar'
     }
 
-    It 'should return the variables for an application' {
+    # Doesn't currently work in BuildMaster.
+    It 'should return the variables for an application' -Skip {
         $app = GivenApplication
         GivenVariable 'GlobalVar' -WithValue 'GlobalSnafu'
         GivenVariable 'AppFubar' -WithValue 'Snafu' -ForApplication $app.Application_Name
@@ -272,17 +278,17 @@ Describe 'Get-BMVariable' {
     }
 
     It 'should ignore WhatIf' {
-        $app = GivenApplication
-        GivenVariable 'AppFubar' -WithValue 'Snafu' -ForApplication $app.Application_Name
+        GivenVariable 'AppFubar' -WithValue 'Snafu'
         $WhatIfPreference = $true
-        WhenGettingVariable 'AppFubar' -ForApplication $app.Application_Name
+        WhenGettingVariable 'AppFubar'
         $WhatIfPreference | Should -BeTrue
         $WhatIfPreference = $false
         ThenVariableReturned @{ 'AppFubar' = 'Snafu' }
         ThenNoErrorWritten
     }
 
-    It 'should return application group''s variable' {
+    # Doesn't currently work in BuildMaster.
+    It 'should return application group''s variable' -Skip {
         GivenApplicationGroup 'fizzbuzz'
         GivenVariable 'GlobalVar' -WithValue 'GlobalSnafu'
         GivenVariable 'AppGroupFubar' -WithValue 'Snafu' -ForApplicationGroup 'fizzbuzz'
@@ -319,41 +325,46 @@ Describe 'Get-BMVariable' {
     }
 
     It 'should URL-encode variable names' {
-        Mock -CommandName 'Invoke-BMRestMethod' -ModuleName 'BuildMasterAutomation'
-        WhenGettingVariable 'EnvironmentFubar' -ForEnvironment 'Get BMVariable'
+        Mock -CommandName 'Invoke-BMRestMethod' -ModuleName 'BuildMasterAutomation' -MockWith {
+            [pscustomobject]@{
+                Name = 'URL Encode Me!'
+                Value = 'ok'
+            }
+        }
+        WhenGettingVariable 'URL Encode Me!'
         Assert-MockCalled -CommandName 'Invoke-BMRestMethod' `
                           -ModuleName 'BuildMasterAutomation' `
-                          -ParameterFilter { $Name -eq 'variables/environment/Get%20BMVariable' }
+                          -ParameterFilter { $Name -eq 'variables/global/URL%20Encode%20Me%21' }
         ThenNoErrorWritten
     }
 
     It 'should write an error when server does not exist' {
         WhenGettingVariable -Named 'Nope' -ForServer 'Nope' -ErrorAction SilentlyContinue
-        ThenError 'server\ was\ not\ found'
+        ThenError ([regex]::Escape('server "Nope" does not exist'))
         ThenNothingReturned
     }
 
     It 'should write an error when server role does not exist' {
         WhenGettingVariable -Named 'Nope' -ForServerRole 'Nope' -ErrorAction SilentlyContinue
-        ThenError 'role\ was\ not\ found'
+        ThenError ([regex]::Escape('server role "Nope" does not exist'))
         ThenNothingReturned
     }
 
     It 'should write an error when environment does not exist' {
         WhenGettingVariable -Named 'Nope' -ForEnvironment 'Nope' -ErrorAction SilentlyContinue
-        ThenError 'environment\ was\ not\ found'
+        ThenError ([regex]::Escape('environment "Nope" does not exist'))
         ThenNothingReturned
     }
 
     It 'should write an error when application does not exist' {
         WhenGettingVariable -Named 'Nope' -ForApplication 'Nope' -ErrorAction SilentlyContinue
-        ThenError 'application\ "Nope"\ does\ not\ exist'
+        ThenError ([regex]::Escape('application "Nope" does not exist'))
         ThenNothingReturned
     }
 
     It 'should write an error when application group does not exist' {
         WhenGettingVariable -Named 'Nope' -ForApplicationGroup 'Nope' -ErrorAction SilentlyContinue
-        ThenError 'application\ group\ "Nope"\ does\ not\ exist'
+        ThenError ([regex]::Escape('application group "Nope" does not exist'))
         ThenNothingReturned
     }