+++ init.ps1
@@ -29,7 +29,14 @@ $InformationPreference = 'Continue'
 
 $runningUnderAppVeyor = (Test-Path -Path 'env:APPVEYOR')
 
+
+# When updating the version, it's a good time to check if bugs in the API have been fixed. Do the following:
+#
+# * Remove the skipped tests in the Get-BMVariable and Remove-BMVariable tests.
 $version = '6.2.33'
+
+
+
 Write-Verbose -Message ('Testing BuildMaster {0}' -f $version)
 $sqlServer = $null
 $installerPath = 'SQL'